+++ audio_hal/aml_audio_stream.h
@@ -334,7 +334,7 @@ struct aml_audio_patch {
     unsigned int dtv_decoder_ready;
     unsigned int input_thread_created;
     unsigned int ouput_thread_created;
-    unsigned int decoder_offset ;
+    uint64_t decoder_offset ;
     unsigned int outlen_after_last_validpts;
     unsigned long last_valid_pts;
     unsigned int first_apts_lookup_over; /*cache audio data before start-play flag*/
