+++ libms12_v24/include/aml_audio_ms12.h
@@ -163,7 +163,7 @@ struct dolby_ms12_desc {
     void *mat_enc_handle;
     char *mat_enc_out_buffer;
     int mat_enc_out_bytes;
-    uint32_t dtv_decoder_offset_base;  /*save the dtv input offset, which is realted with PTS*/
+    uint64_t dtv_decoder_offset_base;  /*save the dtv input offset, which is realted with PTS*/
     /* MAT Encoder inside ms12, end */
     /* For DAP multi output except the stereo output */
     bool tv_tuning_flag;
